requirements.txt
Flask==2.2.1  
Flask-SQLAlchemy==3.0.3
authlib==0.15.4
requests==2.31.0
mysql-connector-python==8.4.0
Werkzeug==2.2.2